add python

+++ src/rust.h
@@ -1,2 +1,2 @@
 #include <stdint.h>
-void store(const uint16_t* vals, size_t len);
+void store(const uint8_t* vals, size_t len);
